backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
alembic==1.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
blinker==1.6.2
greenlet==2.0.2
typing_extensions==4.8.0
APScheduler==3.10.4
stripe==6.6.0
weasyprint==60.1
Pillow==10.0.1
reportlab==4.0.5
PyPDF2==3.0.1 